Add: Dockerised and added README for API Docs
