+++ chatup-nodejs/.gitignore
@@ -1,3 +1,5 @@
 node_modules
 # Keep environment variables out of version control
 .env
+
+public
