+++ chatup-nodejs/package.json
@@ -29,6 +29,7 @@
     "dotenv": "^16.4.5",
     "jest": "^29.7.0",
     "nodemon": "^3.1.7",
+    "path": "^0.12.7",
     "supertest": "^7.0.0"
   }
 }
