+++ docker-compose.yml
@@ -5,7 +5,7 @@ services:
       context: ./chatup-react
       dockerfile: Dockerfile
     ports:
-      - "3000:3000" # React dev server
+      - "5173:5173" # React dev server
     volumes:
       - ./chatup-react:/frontend
       - /frontend/node_modules # Prevents node_modules on host from overwriting container's node_modules
@@ -35,7 +35,7 @@ services:
       - app-network
 
   mongodb:
-    image: mongo:6.0
+    image: mongo:latest
     restart: always
     ports:
       - "27017:27017"
